기본 패키지 구조 설계 및 엔티티, 도메인 작성
