+++ docs/db/erd.md
@@ -51,7 +51,7 @@ CREATE TABLE `SEAT` (
     `seat_no` INT NOT NULL COMMENT '좌석 번호 (1-50)',
     `price` DECIMAL(8,0) NOT NULL COMMENT '좌석 가격',
     `seat_class` ENUM('VIP', 'PREMIUM', 'NORMAL') NOT NULL COMMENT '좌석 등급',
-    `status` ENUM('AVAILABLE', 'RESERVED', 'BLOCKED') NOT NULL DEFAULT 'AVAILABLE' COMMENT '좌석 상태',
+    `status` ENUM('AVAILABLE', 'RESERVED', 'ASSIGNED') NOT NULL DEFAULT 'AVAILABLE' COMMENT '좌석 상태',
     `created_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '생성일시',
     `updated_at` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '수정일시',
     PRIMARY KEY (`id`),
