--- src/main/java/kr/hhplus/be/server/ServerApplication.java
@@ -1,13 +1,0 @@
-package kr.hhplus.be.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ServerApplication.class, args);
-	}
-
-}
