--- src/main/java/kr/hhplus/be/server/domain/concert/model/SeatClass.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.concert.model;
-
-public enum SeatClass {
-	VIP, PREMIUM, NORMAL
-}
