--- src/main/java/kr/hhplus/be/server/domain/concert/model/SeatStatus.java
@@ -1,16 +1,0 @@
-package kr.hhplus.be.server.domain.concert.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public enum SeatStatus {
-	AVAILABLE("예약가능"),
-	RESERVED("예약됨"),
-	ASSIGNED("배정됨")
-	;
-
-	private final String message;
-
-}
