--- src/main/java/kr/hhplus/be/server/domain/payment/model/PaymentStatus.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.payment.model;
-
-public enum PaymentStatus {
-	PENDING, SUCCESS, FAILED, CANCELLED
-}
