--- src/main/java/kr/hhplus/be/server/domain/reservation/model/ReservationStatus.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.reservation.model;
-
-public enum ReservationStatus {
-	PENDING, SUCCESS, FAILED, CANCELLED
-}
