+++ src/main/java/kr/hhplus/be/server/entity/concert/Concert.java
@@ -1,0 +1,13 @@
+package kr.hhplus.be.server.entity.concert;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+public record Concert(
+	UUID id,
+	String title,
+	String artist,
+	LocalDateTime createdAt,
+	LocalDateTime updatedAt
+) {
+}
