+++ src/main/java/kr/hhplus/be/server/entity/concert/ConcertDate.java
@@ -1,0 +1,14 @@
+package kr.hhplus.be.server.entity.concert;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+public record ConcertDate(
+	UUID id,
+	UUID concertId,
+	LocalDateTime date,
+	LocalDateTime deadline,
+	LocalDateTime createdAt,
+	LocalDateTime updatedAt
+) {
+}
