+++ src/main/java/kr/hhplus/be/server/entity/concert/SeatClass.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.entity.concert;
+
+public enum SeatClass {
+	VIP, PREMIUM, NORMAL
+}
