+++ src/main/java/kr/hhplus/be/server/entity/concert/SeatStatus.java
@@ -1,0 +1,16 @@
+package kr.hhplus.be.server.entity.concert;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum SeatStatus {
+	AVAILABLE("예약가능"),
+	RESERVED("예약됨"),
+	ASSIGNED("배정됨")
+	;
+
+	private final String message;
+
+}
