+++ src/main/java/kr/hhplus/be/server/entity/payment/Payment.java
@@ -1,0 +1,17 @@
+package kr.hhplus.be.server.entity.payment;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+public record Payment(
+	UUID id,
+	UUID userId,
+	UUID reservationId,
+	BigDecimal amount,
+	PaymentStatus status,
+	String failureReason,
+	LocalDateTime createdAt,
+	LocalDateTime updatedAt
+) {
+}
