+++ src/main/java/kr/hhplus/be/server/entity/payment/PaymentStatus.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.entity.payment;
+
+public enum PaymentStatus {
+	PENDING, SUCCESS, FAILED, CANCELLED
+}
