+++ src/main/java/kr/hhplus/be/server/entity/reservation/ReservationStatus.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.entity.reservation;
+
+public enum ReservationStatus {
+	PENDING, SUCCESS, FAILED, CANCELLED
+}
