+++ src/main/java/kr/hhplus/be/server/entity/user/User.java
@@ -1,0 +1,17 @@
+package kr.hhplus.be.server.entity.user;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+import lombok.Builder;
+
+
+@Builder
+public record User (
+	UUID id,
+	BigDecimal amount,
+	LocalDateTime createdAt,
+	LocalDateTime updatedAt
+) {
+}
