+++ src/main/java/kr/hhplus/be/server/framework/ServerApplication.java
@@ -1,0 +1,13 @@
+package kr.hhplus.be.server.framework;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ServerApplication.class, args);
+	}
+
+}
