+++ src/main/java/kr/hhplus/be/server/interfaces/api/concert/concert/dto/request/QueueTokenRequest.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.interfaces.api.concert.concert.dto.request;
+
+import java.util.UUID;
+
+public record QueueTokenRequest (
+	UUID userId
+) {
+}
