+++ src/main/java/kr/hhplus/be/server/interfaces/api/concert/concert/dto/request/ReservationRequest.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.interfaces.api.concert.concert.dto.request;
+
+import java.util.UUID;
+
+public record ReservationRequest(
+	UUID concertDateId,
+	UUID seatId
+) {
+}
