--- src/main/java/kr/hhplus/be/server/interfaces/api/concert/dto/request/QueueTokenRequest.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.interfaces.api.concert.dto.request;
-
-import java.util.UUID;
-
-public record QueueTokenRequest (
-	UUID userId
-) {
-}
