+++ src/main/java/kr/hhplus/be/server/interfaces/api/payment/PaymentController.java
@@ -17,7 +17,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import kr.hhplus.be.server.domain.payment.model.PaymentStatus;
+import kr.hhplus.be.server.entity.payment.PaymentStatus;
 import kr.hhplus.be.server.interfaces.api.payment.dto.response.PaymentResponse;
 
 @RestController
