+++ src/main/java/kr/hhplus/be/server/interfaces/api/payment/dto/response/PaymentResponse.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 import java.util.UUID;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.hhplus.be.server.domain.payment.model.PaymentStatus;
+import kr.hhplus.be.server.entity.payment.PaymentStatus;
 
 public record PaymentResponse(
 	@Schema(description = "결제 ID")
