+++ src/main/java/kr/hhplus/be/server/interfaces/gateway/repository/payment/PaymentEntity.java
@@ -1,0 +1,64 @@
+package kr.hhplus.be.server.interfaces.gateway.repository.payment;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.JdbcTypeCode;
+import org.hibernate.type.SqlTypes;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import kr.hhplus.be.server.entity.payment.PaymentStatus;
+import kr.hhplus.be.server.interfaces.gateway.repository.BaseTimeEntity;
+import kr.hhplus.be.server.interfaces.gateway.repository.reservation.ReservationEntity;
+import kr.hhplus.be.server.interfaces.gateway.repository.user.UserEntity;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "PAYMENT")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
+@Builder
+public class PaymentEntity extends BaseTimeEntity {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.UUID)
+	@JdbcTypeCode(SqlTypes.VARCHAR)
+	@Column(name = "id", length = 36)
+	private String id;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id", nullable = false)
+	private UserEntity user;
+
+	@OneToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "reservation_id", nullable = false)
+	private ReservationEntity reservation;
+
+	@Column(name = "amount", precision = 10, nullable = false)
+	private BigDecimal amount;
+
+	@Enumerated(EnumType.STRING)
+	@Column(name = "status", nullable = false)
+	@ColumnDefault("'PENDING'")
+	private PaymentStatus status;
+
+	@Column(name = "failure_reason")
+	private String failureReason;
+}
