+++ src/main/java/kr/hhplus/be/server/interfaces/gateway/repository/reservation/ReservationEntity.java
@@ -1,0 +1,57 @@
+package kr.hhplus.be.server.interfaces.gateway.repository.reservation;
+
+import java.util.UUID;
+
+import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.JdbcTypeCode;
+import org.hibernate.type.SqlTypes;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import kr.hhplus.be.server.entity.reservation.ReservationStatus;
+import kr.hhplus.be.server.interfaces.gateway.repository.BaseTimeEntity;
+import kr.hhplus.be.server.interfaces.gateway.repository.concert.SeatEntity;
+import kr.hhplus.be.server.interfaces.gateway.repository.user.UserEntity;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "RESERVATION")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
+@Builder
+public class ReservationEntity extends BaseTimeEntity {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.UUID)
+	@JdbcTypeCode(SqlTypes.VARCHAR)
+	@Column(name = "id", length = 36)
+	private String id;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id", nullable = false)
+	private UserEntity user;
+
+	@OneToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "seat_id", nullable = false)
+	private SeatEntity seat;
+
+	@Enumerated(EnumType.STRING)
+	@Column(name = "status", nullable = false)
+	@ColumnDefault("'PENDING'")
+	private ReservationStatus status;
+}
