+++ src/main/java/kr/hhplus/be/server/interfaces/gateway/repository/user/UserRepository.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.interfaces.gateway.repository.user;
+
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserEntity, UUID> {
+}
