+++ src/main/java/kr/hhplus/be/server/usecase/user/UserService.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.usecase.user;
+
+public class UserService {
+
+}
